Working version of the refactored simulation processor. Functions for loading external effects have been moved to utils.
