+++ brainspy/processors/simulation/processor.py
@@ -46,6 +46,7 @@ import collections
 import numpy as np
 from torch import nn
 
+from brainspy.utils.electrodes import set_effects_from_dict
 from brainspy.utils.pytorch import TorchUtils
 from brainspy.processors.simulation.noise.noise import get_noise
 from brainspy.processors.simulation.model import NeuralNetworkModel
@@ -70,7 +71,6 @@ class SurrogateModel(nn.Module):
     def __init__(
         self,
         model_structure: dict,
-        electrode_info: dict,
         model_state_dict: collections.OrderedDict = None,
     ):
         """
@@ -106,127 +106,9 @@ class SurrogateModel(nn.Module):
         self.model = NeuralNetworkModel(model_structure)
         if model_state_dict is not None:
             self.model.load_state_dict(model_state_dict)
-        self.electrode_info = electrode_info
-        self.set_voltage_ranges("default")
-        self.set_amplification("default")
-
-    # TODO: Add description of this method
-    def set_effects_from_dict(self, configs):
-        self.set_effects(
-            self._get_key(configs, "voltage_ranges"),
-            self._get_key(configs, "amplification"),
-            self._get_key(configs, "output_clipping"),
-            self._get_key(configs, "noise"),
-        )
-
-    # TODO: Add description of this method
-    def _get_key(self, configs, effect_key):
-        if effect_key in configs:
-            return configs[effect_key]
-        if effect_key != "noise":
-            return "default"
-        return None
-
-    def set_effects(
-        self,
-        voltage_ranges="default",
-        amplification="default",
-        output_clipping="default",
-        noise_configs=None,
-    ):
-        """
-        Set the amplification, output clipping and noise of the processor.
-        Amplification and output clipping are explained in their respective
-        methods. Noise is an error which is superimposed on the output of the
-        network to give it an element of randomness.
-
-        Order of effects: amplification - noise - output clipping
-
-        Example
-        -------
-        >>> smg = SurrogateModel("model.pt")
-        >>> smg.set_effects(amplification=2.0,
-                            output_clipping="default",
-                            noise=None)
-
-        Parameters
-        ----------
-        voltage_ranges:
-            Voltage ranges of the activation electrodes. Can be a value or 'default'.
-        amplification
-            The amplification of the processor. Can be None, a value, or
-            'default'. By default None.
-        output_clipping
-            The output clipping of the processor. Can be None, a value, or
-            'default'. By default None.
-        noise
-            The noise of the processor. Can be None, a string determining
-            the type of noise and some args. By default None.
-        """
-        # Warning, this function used to be called form the init using a
-        # configs file. Now it is called externally. To be changed where it
-        # corresponds in bspy tasks.
-        self.set_amplification(amplification)
-        self.set_output_clipping(output_clipping)
-        self.noise = get_noise(noise_configs)
-
-    def set_voltage_ranges(self, value):
-        # TODO: Document this function.
-        if value is not None and value == "default":
-            self.voltage_ranges = TorchUtils.format(
-                self.electrode_info["activation_electrodes"]["control_voltages"]
-            )
 
     def get_voltage_ranges(self):
         return self.voltage_ranges
-
-    def set_amplification(self, value):
-        """
-        Set the amplification of the processor. The amplificaiton is what the
-        output of the neural network is multiplied with after the forward pass.
-        Can be None, a value, or 'default', by default None.
-        None will not use amplification, a value will set the amplification
-        to that value, and the string 'default' will take the data from the
-        info dictionary.
-
-        This method is called through the "set_effects" method.
-
-        Parameters
-        ----------
-        value : None or double or str
-            The value of the amplification (None, a value or 'default').
-        """
-        if value is not None and value == "default":
-            self.amplification = TorchUtils.format(
-                self.electrode_info["output_electrodes"]["amplification"]
-            )
-        else:
-            self.amplification = value
-
-    def set_output_clipping(self, value: torch.Tensor):
-        """
-        Set the output clipping of the processor. Output clipping means to
-        clip the output to a certain range. Any output above that range will
-        be replaced with the maximum and any output below will be set to the
-        minimum.
-        Can be None, a value, or 'default'.
-        None will not use clipping, a value will set the clipping to that
-        value, and the string 'default' will take the data from the info
-        dictionary.
-
-        This method is called through the "set_effects" method.
-
-        Parameters
-        ----------
-        value : None or double or str
-            The value of the output clipping (None, a value or 'default').
-        """
-        if value is not None and value == "default":
-            self.output_clipping = TorchUtils.format(
-                self.electrode_info["output_electrodes"]["clipping_value"]
-            )
-        else:
-            self.output_clipping = value
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         """
@@ -314,14 +196,3 @@ class SurrogateModel(nn.Module):
             False
         """
         return False
-
-    def get_electrode_no(self):
-        """
-        Get the number of electrodes of the processor.
-
-        Returns
-        -------
-        int
-            The number of electrodes of the processor.
-        """
-        return len(self.simulation_params["electrode_no"])
